Fix language-at-point.
If the current templ file has no script components, the javascript parser
has no ranges assigned... Which means the parser is being applied to and
tried on the entire file, which means treesit-parser-range-on always returns
a non-nil range for javascript. So, first explicitly check whether Javascript
is being restricted to ranges, and only listen to it if yes.