Fix subparser handling.
The simple range rules in Emacs 29 do not correctly handle the case
where a language matches none of the buffer: instead of setting the
relevant parser to process nothing, it's set to parse the entire file.
This causes unfortunate mixups in at least font-locking, and possibly
other places.

So, replace the rule-based range setting with a function that always
updates the entire buffer with correct javascript ranges. This is
more wasteful than the rule-based incremental updates, but it's
good enough for now until emacs figures things out.